+++ constrainedfilefield/models.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Placeholder to make it work as an app.
-"""