+++ constrainedfilefield/tests/models.py
@@ -1,6 +1,6 @@
 from django.db import models
 
-from ..fields import ConstrainedFileField
+from constrainedfilefield.fields import ConstrainedFileField
 
 
 class TestModel(models.Model):
@@ -10,6 +10,16 @@ class TestModel(models.Model):
         upload_to='testfile',
         content_types=['image/png'],
         max_upload_size=10240)
+
+
+class TestModelJs(models.Model):
+    the_file = ConstrainedFileField(
+        null=True,
+        blank=True,
+        upload_to='testfile',
+        content_types=['image/png'],
+        max_upload_size=10240,
+        js_checker=True)
 
 
 class TestModelNoValidate(models.Model):